Added validation to service layer.

+++ src/main/java/com/sergeev/studapp/service/DisciplineService.java
@@ -7,7 +7,7 @@ import com.sergeev.studapp.model.Discipline;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 public class DisciplineService {
@@ -15,8 +15,15 @@ public class DisciplineService {
     private static final Logger LOG = LoggerFactory.getLogger(DisciplineService.class);
     private static final DisciplineDao DISCIPLINE_DAO = DaoFactory.getDaoFactory(DaoFactory.POSTGRES).getDisciplineDao();
 
-    public static Discipline create(String title){
-        Discipline discipline = new Discipline();
+    private static Discipline discipline;
+    private static List<Discipline> disciplines;
+
+    public static Discipline create(String title) throws ApplicationException {
+        if (!checkTitle(title)){
+            throw new ApplicationException("Bad parameters.");
+        }
+
+        discipline = new Discipline();
         discipline.setTitle(title);
 
         try {
@@ -28,50 +35,63 @@ public class DisciplineService {
         return discipline;
     }
 
-    public static Discipline read(String id) {
-        Discipline discipline = null;
+    public static Discipline read(String id) throws ApplicationException {
 
         try {
             discipline = DISCIPLINE_DAO.getById(id);
         } catch (PersistentException e) {
-            e.printStackTrace();
+            throw new ApplicationException("Discipline not found.", e);
         }
 
         return discipline;
     }
 
     public static List<Discipline> readAll(){
-        List<Discipline> disciplines = new ArrayList<>();
 
         try {
             disciplines = DISCIPLINE_DAO.getAll();
         } catch (PersistentException e) {
-            e.printStackTrace();
+            disciplines  = Collections.emptyList();
         }
 
         return disciplines;
     }
 
-    public static Discipline update(String title, String id){
-        Discipline discipline = new Discipline();
+    public static Discipline update(String title, String id) throws ApplicationException {
+        if (id == null || id.isEmpty() || !checkTitle(title)) {
+            throw new ApplicationException("Bad parameters.");
+        }
+
+        discipline = new Discipline();
         discipline.setId(id);
         discipline.setTitle(title);
 
         try {
             DISCIPLINE_DAO.update(discipline);
         } catch (PersistentException e) {
-            e.printStackTrace();
+            throw new ApplicationException("Cannot update discipline.", e);
         }
 
         return discipline;
     }
 
-    public static void delete(String id) {
+    public static void delete(String id) throws ApplicationException {
+        if (id == null || id.isEmpty()){
+            throw new ApplicationException("Bad parameters.");
+        }
+
         try {
             DISCIPLINE_DAO.delete(id);
         } catch (PersistentException e) {
-            e.printStackTrace();
+            throw new ApplicationException("Cannot delete group, because group not found.", e);
         }
     }
 
+    private static boolean checkTitle(String title) {
+        if (title == null || title.isEmpty()) {
+            return false;
+        }
+        String expression = "(?u)^\\p{Lu}\\w+$";
+        return title.matches(expression);
+    }
 }
